RAD-232 Move dicom codes to own package
