+++ api/src/main/java/org/openmrs/module/radiology/dicom/code/PerformedProcedureStepStatus.java
@@ -7,7 +7,7 @@
  * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS 
  * graphic logo is a trademark of OpenMRS Inc.
  */
-package org.openmrs.module.radiology;
+package org.openmrs.module.radiology.dicom.code;
 
 /**
  * <p>
