+++ api/src/main/java/org/openmrs/module/radiology/study/RadiologyStudy.java
@@ -13,8 +13,8 @@ import java.lang.reflect.Field;
 
 import org.openmrs.module.radiology.Modality;
 import org.openmrs.module.radiology.MwlStatus;
-import org.openmrs.module.radiology.PerformedProcedureStepStatus;
-import org.openmrs.module.radiology.ScheduledProcedureStepStatus;
+import org.openmrs.module.radiology.dicom.code.PerformedProcedureStepStatus;
+import org.openmrs.module.radiology.dicom.code.ScheduledProcedureStepStatus;
 import org.openmrs.module.radiology.order.RadiologyOrder;
 
 /**
