+++ api/src/main/resources/RadiologyStudy.hbm.xml
@@ -2,7 +2,8 @@
 <!DOCTYPE hibernate-mapping PUBLIC "-//Hibernate/Hibernate Mapping DTD 3.0//EN" "http://hibernate.sourceforge.net/hibernate-mapping-3.0.dtd">
 
 <hibernate-mapping package="org.openmrs.module.radiology">
-	<class name="org.openmrs.module.radiology.study.RadiologyStudy" table="radiology_study">
+	<class name="org.openmrs.module.radiology.study.RadiologyStudy"
+		table="radiology_study">
 		<id name="studyId" type="int" column="study_id" unsaved-value="0">
 			<generator class="native">
 				<param name="sequence">radiology_study_study_id_seq</param>
@@ -16,13 +17,13 @@
 		<property name="scheduledStatus" column="scheduled_status"
 			length="9">
 			<type name="org.openmrs.util.HibernateEnumType">
-				<param name="enumClassName">org.openmrs.module.radiology.ScheduledProcedureStepStatus</param>
+				<param name="enumClassName">org.openmrs.module.radiology.dicom.code.ScheduledProcedureStepStatus</param>
 			</type>
 		</property>
 		<property name="performedStatus" column="performed_status"
 			length="12">
 			<type name="org.openmrs.util.HibernateEnumType">
-				<param name="enumClassName">org.openmrs.module.radiology.PerformedProcedureStepStatus</param>
+				<param name="enumClassName">org.openmrs.module.radiology.dicom.code.PerformedProcedureStepStatus</param>
 			</type>
 		</property>
 		<property name="modality" column="modality" length="16"
