+++ api/src/test/java/org/openmrs/module/radiology/dicom/code/ScheduledProcedureStepStatusTest.java
@@ -1,4 +1,4 @@
-package org.openmrs.module.radiology;
+package org.openmrs.module.radiology.dicom.code;
 
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertThat;
